+++ main.py
@@ -7,7 +7,8 @@ from ipaddress import IPv4Address
 from os import mkdir
 from random import shuffle
 from shutil import rmtree
-from typing import Any, Dict, Iterable, Optional, Tuple
+from time import perf_counter
+from typing import Callable, Dict, Iterable, List, Optional, Tuple, Union
 
 from aiohttp import ClientSession
 from aiohttp_socks import ProxyConnector
@@ -17,12 +18,53 @@ from maxminddb.reader import Reader
 import config
 
 
+class Proxy:
+    def __init__(self, socket_address: str) -> None:
+        self.socket_address = socket_address
+        self._ip = socket_address.split(":")[0]
+        self.exit_node: Optional[str] = None
+        self.is_anonymous: Optional[bool] = None
+        self.geolocation = "::None::None::None"
+        self.timeout: float = float("inf")
+
+    def set_anonymity(self) -> None:
+        self.is_anonymous = self._ip != self.exit_node
+
+    def set_geolocation(self, reader: Reader) -> None:
+        geolocation = reader.get(self._ip)
+        if not isinstance(geolocation, dict):
+            return
+        country = geolocation.get("country")
+        if country:
+            country = country["names"]["en"]
+        else:
+            country = geolocation.get("continent")
+            if country:
+                country = country["names"]["en"]
+        region = geolocation.get("subdivisions")
+        if region:
+            region = region[0]["names"]["en"]
+        city = geolocation.get("city")
+        if city:
+            city = city["names"]["en"]
+        self.geolocation = f"::{country}::{region}::{city}"
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, Proxy):
+            return NotImplemented
+        return self.socket_address == other.socket_address
+
+    def __hash__(self) -> int:
+        return hash(("socket_address", self.socket_address))
+
+
 class ProxyScraperChecker:
     def __init__(
         self,
         *,
+        timeout: float = 10,
         max_connections: int = 950,
-        timeout: float = 5,
+        sort_by_speed: bool = True,
         geolite2_city_mmdb: Optional[str] = None,
         ip_service: str = "https://checkip.amazonaws.com",
         http_sources: Optional[Iterable[str]] = None,
@@ -41,6 +83,7 @@ class ProxyScraperChecker:
         """
         self.sem = asyncio.Semaphore(max_connections)
         self.IP_SERVICE = ip_service.strip()
+        self.SORT_BY_SPEED = sort_by_speed
         self.TIMEOUT = timeout
         self.MMDB = geolite2_city_mmdb
         self.SOURCES = {
@@ -54,41 +97,10 @@ class ProxyScraperChecker:
             )
             if sources
         }
-        self.proxies: Dict[str, Dict[str, Optional[str]]] = {
-            proto: {} for proto in self.SOURCES
+        self.proxies: Dict[str, List[Proxy]] = {
+            proto: [] for proto in self.SOURCES
         }
         self.proxies_count = {proto: 0 for proto in self.SOURCES}
-
-    @staticmethod
-    def get_geolocation(ip: Optional[str], reader: Reader) -> str:
-        """Get proxy's geolocation.
-
-        Args:
-            ip (str): Proxy's ip.
-            reader (Reader): mmdb Reader instance.
-
-        Returns:
-            str: ::Country Name::Region::City
-        """
-        if not ip:
-            return "::None::None::None"
-        geolocation = reader.get(ip)
-        if not isinstance(geolocation, dict):
-            return "::None::None::None"
-        country = geolocation.get("country")
-        if country:
-            country = country["names"]["en"]
-        else:
-            country = geolocation.get("continent")
-            if country:
-                country = country["names"]["en"]
-        region = geolocation.get("subdivisions")
-        if region:
-            region = region[0]["names"]["en"]
-        city = geolocation.get("city")
-        if city:
-            city = city["names"]["en"]
-        return f"::{country}::{region}::{city}"
 
     async def fetch_source(
         self, session: ClientSession, source: str, proto: str
@@ -117,26 +129,30 @@ class ProxyScraperChecker:
                         IPv4Address(proxy.split(":")[0])
                     except Exception:
                         continue
-                    self.proxies[proto][proxy] = None
+                    self.proxies[proto].append(Proxy(proxy))
             else:
                 logging.error("%s status code: %s", source, status)
 
-    async def check_proxy(self, proxy: str, proto: str) -> None:
+    async def check_proxy(self, proxy: Proxy, proto: str) -> None:
         """Check proxy validity.
 
         Args:
-            proxy (str): ip:port.
+            proxy (Proxy): ip:port.
             proto (str): http/socks4/socks5.
         """
+        start = perf_counter()
         try:
             async with self.sem:
                 async with ClientSession(
-                    connector=ProxyConnector.from_url(f"{proto}://{proxy}")
+                    connector=ProxyConnector.from_url(
+                        f"{proto}://{proxy.socket_address}"
+                    )
                 ) as session:
                     async with session.get(
                         self.IP_SERVICE, timeout=self.TIMEOUT
                     ) as r:
                         exit_node = await r.text(encoding="utf-8")
+            proxy.timeout = perf_counter() - start
             exit_node = exit_node.strip()
             IPv4Address(exit_node)
         except Exception as e:
@@ -145,9 +161,10 @@ class ProxyScraperChecker:
             if isinstance(e, OSError) and e.errno == 24:
                 logging.error("Please, set MAX_CONNECTIONS to lower value.")
 
-            self.proxies[proto].pop(proxy)
+            self.proxies[proto].remove(proxy)
         else:
-            self.proxies[proto][proxy] = exit_node
+            proxy.exit_node = exit_node
+            proxy.set_anonymity()
 
     async def fetch_all_sources(self) -> None:
         """Get proxies from sources."""
@@ -159,6 +176,12 @@ class ProxyScraperChecker:
                 for source in sources
             )
             await asyncio.gather(*coroutines)
+
+        # Remove duplicates
+        for proto in self.proxies:
+            self.proxies[proto] = list(frozenset(self.proxies[proto]))
+
+        # Remember total count so we could print it in the table
         for proto, proxies in self.proxies.items():
             self.proxies_count[proto] = len(proxies)
 
@@ -173,11 +196,18 @@ class ProxyScraperChecker:
         shuffle(coroutines)
         await asyncio.gather(*coroutines)
 
+    def set_geolocation(self) -> None:
+        if not self.MMDB:
+            return
+        with open_database(self.MMDB) as reader:
+            for proxies in self.proxies.values():
+                for proxy in proxies:
+                    proxy.set_geolocation(reader)
+
     def sort_proxies(self) -> None:
-        self.proxies = {
-            proto: dict(sorted(proxies.items(), key=self._get_sorting_key))
-            for proto, proxies in self.proxies.items()
-        }
+        key = self._sorting_key
+        for proto in self.proxies:
+            self.proxies[proto].sort(key=key)
 
     def save_proxies(self) -> None:
         """Delete old proxies and save new ones."""
@@ -196,20 +226,14 @@ class ProxyScraperChecker:
         for dir in dirs_to_create:
             mkdir(dir)
 
-        anon_proxies = {
-            proto: {
-                proxy: exit_node
-                for proxy, exit_node in proxies.items()
-                if exit_node != proxy.split(":")[0]
-            }
-            for proto, proxies in self.proxies.items()
-        }
         # proxies and proxies_anonymous folders
         for proto, proxies in self.proxies.items():
-            text = "\n".join(proxies)
-            with open(f"proxies/{proto}", "w", encoding="utf-8") as f:
+            text = "\n".join(proxy.socket_address for proxy in proxies)
+            with open(f"proxies/{proto}.txt", "w", encoding="utf-8") as f:
                 f.write(text)
-            anon_text = "\n".join(anon_proxies[proto])
+            anon_text = "\n".join(
+                proxy.socket_address for proxy in proxies if proxy.is_anonymous
+            )
             with open(
                 f"proxies_anonymous/{proto}.txt", "w", encoding="utf-8"
             ) as f:
@@ -218,25 +242,20 @@ class ProxyScraperChecker:
         # proxies_geolocation and proxies_geolocation_anonymous folders
         if not self.MMDB:
             return
-        with open_database(self.MMDB) as reader:
-            geolocations = {
-                proto: {
-                    proxy: self.get_geolocation(exit_node, reader)
-                    for proxy, exit_node in proxies.items()
-                }
-                for proto, proxies in self.proxies.items()
-            }
+        self.set_geolocation()
         for proto, proxies in self.proxies.items():
             text = "\n".join(
-                f"{proxy}{geolocations[proto][proxy]}" for proxy in proxies
+                f"{proxy.socket_address}{proxy.geolocation}"
+                for proxy in proxies
             )
             with open(
                 f"proxies_geolocation/{proto}.txt", "w", encoding="utf-8"
             ) as f:
                 f.write(text)
             anon_text = "\n".join(
-                f"{proxy}{geolocations[proto][proxy]}"
-                for proxy in anon_proxies[proto]
+                f"{proxy.socket_address}{proxy.geolocation}"
+                for proxy in proxies
+                if proxy.is_anonymous
             )
             with open(
                 f"proxies_geolocation_anonymous/{proto}.txt",
@@ -261,10 +280,13 @@ class ProxyScraperChecker:
         )
         logging.info("Thank you for using proxy-scraper-checker :)")
 
-    @staticmethod
-    def _get_sorting_key(x: Tuple[str, Any]) -> Tuple[int, ...]:
-        octets = x[0].replace(":", ".").split(".")
-        return tuple(map(int, octets))
+    @property
+    def _sorting_key(self) -> Callable[[Proxy], Union[float, Tuple[int, ...]]]:
+        if self.SORT_BY_SPEED:
+            return lambda proxy: tuple(
+                map(int, proxy.socket_address.replace(":", ".").split("."))
+            )
+        return lambda proxy: proxy.timeout
 
 
 async def main() -> None:
@@ -274,8 +296,9 @@ async def main() -> None:
         datefmt="%Y-%m-%d %H:%M:%S",
     )
     await ProxyScraperChecker(
-        max_connections=config.MAX_CONNECTIONS,
         timeout=config.TIMEOUT,
+        max_connections=config.MAX_CONNECTIONS,
+        sort_by_speed=config.SORT_BY_SPEED,
         geolite2_city_mmdb="GeoLite2-City.mmdb"
         if config.GEOLOCATION
         else None,